requirements.txt
Flask
Gunicorn
Flask-SQLAlchemy
Flask-Login==0.6.3
oauthlib==3.2.2
requests==2.31.0
python-dotenv
azure-ai-formrecognizer==3.3.0
openai>=1.0.0
azure-identity==1.15.0
Pillow==9.5.0
APScheduler==3.10.4
pytz==2023.3.post1
